+++ src/main/java/com/fasterxml/jackson/core/json/ReaderBasedJsonParser.java
@@ -1405,7 +1405,7 @@ public class ReaderBasedJsonParser
                 _inputPtr = startPtr;
                 return _parseNumber2(false, startPtr);
             }
-            ch = (int) _inputBuffer[ptr++];
+            ch = _inputBuffer[ptr++];
             if (ch < INT_0 || ch > INT_9) {
                 break int_loop;
             }
@@ -1440,7 +1440,7 @@ public class ReaderBasedJsonParser
                 if (ptr >= inputLen) {
                     return _parseNumber2(neg, startPtr);
                 }
-                ch = (int) _inputBuffer[ptr++];
+                ch = _inputBuffer[ptr++];
                 if (ch < INT_0 || ch > INT_9) {
                     break fract_loop;
                 }
@@ -1460,13 +1460,13 @@ public class ReaderBasedJsonParser
                 return _parseNumber2(neg, startPtr);
             }
             // Sign indicator?
-            ch = (int) _inputBuffer[ptr++];
+            ch = _inputBuffer[ptr++];
             if (ch == INT_MINUS || ch == INT_PLUS) { // yup, skip for now
                 if (ptr >= inputLen) {
                     _inputPtr = startPtr;
                     return _parseNumber2(neg, startPtr);
                 }
-                ch = (int) _inputBuffer[ptr++];
+                ch = _inputBuffer[ptr++];
             }
             while (ch <= INT_9 && ch >= INT_0) {
                 ++expLen;
@@ -1474,7 +1474,7 @@ public class ReaderBasedJsonParser
                     _inputPtr = startPtr;
                     return _parseNumber2(neg, startPtr);
                 }
-                ch = (int) _inputBuffer[ptr++];
+                ch = _inputBuffer[ptr++];
             }
             // must be followed by sequence of ints, one minimum
             if (expLen == 0) {
@@ -1525,7 +1525,7 @@ public class ReaderBasedJsonParser
             if (ptr >= inputEnd) {
                 return _parseNumber2(negative, startPtr);
             }
-            ch = (int) _inputBuffer[ptr++];
+            ch = _inputBuffer[ptr++];
             if (ch < INT_0 || ch > INT_9) {
                 break int_loop;
             }
@@ -1876,7 +1876,7 @@ public class ReaderBasedJsonParser
                 }
             }
             char c = _inputBuffer[_inputPtr++];
-            int i = (int) c;
+            int i = c;
             if (i <= INT_BACKSLASH) {
                 if (i == INT_BACKSLASH) {
                     /* Although chars outside of BMP are to be escaped as
@@ -2095,7 +2095,7 @@ public class ReaderBasedJsonParser
                 }
             }
             char c = _inputBuffer[_inputPtr++];
-            int i = (int) c;
+            int i = c;
             if (i <= '\\') {
                 if (i == '\\') {
                     // Although chars outside of BMP are to be escaped as
@@ -2137,7 +2137,7 @@ public class ReaderBasedJsonParser
                 }
             }
             char c = _inputBuffer[_inputPtr];
-            int i = (int) c;
+            int i = c;
             if (i < maxCode) {
                 if (codes[i] != 0) {
                     break;
@@ -2215,7 +2215,7 @@ public class ReaderBasedJsonParser
                 }
             }
             char c = _inputBuffer[_inputPtr++];
-            int i = (int) c;
+            int i = c;
             if (i < maxCode && codes[i] != 0) {
                 if (i == INT_QUOTE) {
                     break;
@@ -2267,7 +2267,7 @@ public class ReaderBasedJsonParser
                 inLen = _inputEnd;
             }
             char c = inBuf[inPtr++];
-            int i = (int) c;
+            int i = c;
             if (i <= INT_BACKSLASH) {
                 if (i == INT_BACKSLASH) {
                     // Although chars outside of BMP are to be escaped as an UTF-16 surrogate pair,
@@ -2324,7 +2324,7 @@ public class ReaderBasedJsonParser
                 return i;
             }
             if (i == INT_SPACE || i == INT_TAB) {
-                i = (int) _inputBuffer[++_inputPtr];
+                i = _inputBuffer[++_inputPtr];
                 if (i > INT_SPACE) {
                     if (i == INT_SLASH || i == INT_HASH) {
                         return _skipColon2(true);
@@ -2348,7 +2348,7 @@ public class ReaderBasedJsonParser
                 return i;
             }
             if (i == INT_SPACE || i == INT_TAB) {
-                i = (int) _inputBuffer[++_inputPtr];
+                i = _inputBuffer[++_inputPtr];
                 if (i > INT_SPACE) {
                     if (i == INT_SLASH || i == INT_HASH) {
                         return _skipColon2(true);
@@ -2365,7 +2365,7 @@ public class ReaderBasedJsonParser
     private final int _skipColon2(boolean gotColon) throws IOException
     {
         while (_inputPtr < _inputEnd || _loadMore()) {
-            int i = (int) _inputBuffer[_inputPtr++];
+            int i = _inputBuffer[_inputPtr++];
             if (i > INT_SPACE) {
                 if (i == INT_SLASH) {
                     _skipComment();
@@ -2404,7 +2404,7 @@ public class ReaderBasedJsonParser
     // Variant called when we know there's at least 4 more bytes available
     private final int _skipColonFast(int ptr) throws IOException
     {
-        int i = (int) _inputBuffer[ptr++];
+        int i = _inputBuffer[ptr++];
         if (i == INT_COLON) { // common case, no leading space
             i = _inputBuffer[ptr++];
             if (i > INT_SPACE) { // nor trailing
@@ -2413,7 +2413,7 @@ public class ReaderBasedJsonParser
                     return i;
                 }
             } else if (i == INT_SPACE || i == INT_TAB) {
-                i = (int) _inputBuffer[ptr++];
+                i = _inputBuffer[ptr++];
                 if (i > INT_SPACE) {
                     if (i != INT_SLASH && i != INT_HASH) {
                         _inputPtr = ptr;
@@ -2436,7 +2436,7 @@ public class ReaderBasedJsonParser
                     return i;
                 }
             } else if (i == INT_SPACE || i == INT_TAB) {
-                i = (int) _inputBuffer[ptr++];
+                i = _inputBuffer[ptr++];
                 if (i > INT_SPACE) {
                     if (i != INT_SLASH && i != INT_HASH) {
                         _inputPtr = ptr;
@@ -2456,7 +2456,7 @@ public class ReaderBasedJsonParser
             _reportUnexpectedChar(i, "was expecting comma to separate "+_parsingContext.typeDesc()+" entries");
         }
         while (_inputPtr < _inputEnd) {
-            i = (int) _inputBuffer[_inputPtr++];
+            i = _inputBuffer[_inputPtr++];
             if (i > INT_SPACE) {
                 if (i == INT_SLASH || i == INT_HASH) {
                     --_inputPtr;
@@ -2481,7 +2481,7 @@ public class ReaderBasedJsonParser
     private final int _skipAfterComma2() throws IOException
     {
         while (_inputPtr < _inputEnd || _loadMore()) {
-            int i = (int) _inputBuffer[_inputPtr++];
+            int i = _inputBuffer[_inputPtr++];
             if (i > INT_SPACE) {
                 if (i == INT_SLASH) {
                     _skipComment();
@@ -2537,7 +2537,7 @@ public class ReaderBasedJsonParser
         }
 
         while (_inputPtr < _inputEnd) {
-            i = (int) _inputBuffer[_inputPtr++];
+            i = _inputBuffer[_inputPtr++];
             if (i > INT_SPACE) {
                 if (i == INT_SLASH || i == INT_HASH) {
                     --_inputPtr;
@@ -2567,7 +2567,7 @@ public class ReaderBasedJsonParser
                     return _eofAsNextChar();
                 }
             }
-            int i = (int) _inputBuffer[_inputPtr++];
+            int i = _inputBuffer[_inputPtr++];
             if (i > INT_SPACE) {
                 if (i == INT_SLASH) {
                     _skipComment();
@@ -2615,7 +2615,7 @@ public class ReaderBasedJsonParser
     {
         // Ok: need the matching '*/'
         while ((_inputPtr < _inputEnd) || _loadMore()) {
-            int i = (int) _inputBuffer[_inputPtr++];
+            int i = _inputBuffer[_inputPtr++];
             if (i <= '*') {
                 if (i == '*') { // end?
                     if ((_inputPtr >= _inputEnd) && !_loadMore()) {
@@ -2655,7 +2655,7 @@ public class ReaderBasedJsonParser
     {
         // Ok: need to find EOF or linefeed
         while ((_inputPtr < _inputEnd) || _loadMore()) {
-            int i = (int) _inputBuffer[_inputPtr++];
+            int i = _inputBuffer[_inputPtr++];
             if (i < INT_SPACE) {
                 if (i == INT_LF) {
                     ++_currInputRow;
@@ -2681,7 +2681,7 @@ public class ReaderBasedJsonParser
         }
         char c = _inputBuffer[_inputPtr++];
 
-        switch ((int) c) {
+        switch (c) {
             // First, ones that are mapped
         case 'b':
             return '\b';
@@ -2715,7 +2715,7 @@ public class ReaderBasedJsonParser
                     _reportInvalidEOF(" in character escape sequence", JsonToken.VALUE_STRING);
                 }
             }
-            int ch = (int) _inputBuffer[_inputPtr++];
+            int ch = _inputBuffer[_inputPtr++];
             int digit = CharTypes.charToHex(ch);
             if (digit < 0) {
                 _reportUnexpectedChar(ch, "expected a hex-digit for character escape sequence");
